fix: Do not generate symlinks for major.minor.patch
Those version files are meant to be regular files and not symlinks.
Pre-release versions such as 5.1.10-beta.0 should not create a symlink
5.1.10.json, because that file name should be reserved for the actual
5.1.10 release.

At the moment, craft refuses to publish a new release because the
registry already contains a version file for the release being
published.

Note that the files under packages/nuget/ being removed refer to version
numbers that have never been released (they do not exist in Nuget, only
the preview versions exist).