fix(aws-serverless): Remove faulty 8.43.0 Node Lambda Layer entry (#171)
emoves the 8.43.0 entry for the Node AWS Lambda layer. The reason for this manual fix is that our 8.43.0 publishing run overwrote the correct lambda layer entry with a v8-specific layer we published afterwards. We've fixed this in the meantime in the SDK's publishing config but currently, docs (and possibly our product<>AWS Lambda integration) list the v8-specific layer as the latest version.