fix(aws-serverless): Remove faulty `8.46.0` and `8.47.0` Node Lambda layer entries
Manually removes the faulty releases that overwrote the node lambda layer with the
suffixed version (`SentryNodeServerlessSDKv8`) and links to the last version
(`8.45.1`) that had a correct entry for the layer (`SentryNodeServerlessSDK`).

Docs and the product AWS Lambda integration are not correctly set up when the
suffixed layer is the latest layer.

Release [8.48.0](https://github.com/getsentry/publish/issues/4819) will fix this again.