fix: Manually restore the v9 node layer (#201)
After `9.40.0` we started publishing `10.0.0-alpha` releases which ended up
overwriting the layer in the release registry for v9.

We manually overwrite the versions here so that the next release gets the proper
v9 layer written in.

This came from a mistunderstanding of how craft handles pre-releases (assumption
was: we publish to the AWS registry but not our release registry).