fix: Manually cache AWS Lambda response to avoid empty responses (#206)
This patch applies a "pragmatic" fix for a problem we observed in prod: Every now and then, our `/aws-lambda-layers` endpoint would return a 200 response with an empty response body. Our running theory is that, because [simple cache](https://cachelib.readthedocs.io/en/stable/simple/) is not thread safe (and should not be used in prod 👀) we run into race conditions where we somehow either cache or retrieve empty response bodies for the lambda route.

This is an attempt to quick-fix and hence also verify our theory, by replacing simple cache with storing the response in memory. Data-wise, this is fine because the `/aws-lambda-layers` endpoint's response doesn't change during a release registry deploy. 

---------

Co-authored-by: Michi Hoffmann <cleptric@users.noreply.github.com>